Don't do operations on a closed `ZipFile` (#68)
When zip file was opened inside the loop and it did not contain the lib
inside of it, the `zipFile` stayed non-`null` for all next iterations. This
can cause next iterations that fail to open zip file to access
previously already closed `zipFile` because `zipFile == null` will never
be true.

Fixed by moving the `zipFile` declaration inside of the loop.